requirements.txt
aspose-words==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
dill==0.3.7
idna==3.4
pyTelegramBotAPI==4.14.0
requests==2.31.0
svg-turtle==0.4.2
svgwrite==1.4.3
telebot==0.0.5
urllib3==2.1.0